Return DELETE count and clean up code
