+++ pg_analytics/src/errors.rs
@@ -84,9 +84,6 @@ pub enum NotSupported {
     #[error("UPDATE is not yet supported for deltalake tables")]
     Update,
 
-    #[error("DELETE is not yet supported for deltalake tables")]
-    Delete,
-
     #[error("Heap and deltalake tables in the same query is not yet supported")]
     MixedTables,
 }
